refactoring & standardization: Moved phase_fitting tools from _panel.py to _phase_fitting.py
